Unbreak the 64 bit build. Review URL: http://codereview.chromium.org/9178011
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10414 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/snapshot.h
@@ -90,7 +90,7 @@ class Snapshot {
 
 // These are the sizes of the spaces that are needed in order to unpack the
 // VM boot snapshot.
-const int kMinimumSpaceSizes[LAST_SPACE + 1] = {
+const intptr_t kMinimumSpaceSizes[LAST_SPACE + 1] = {
   0,           // New space.
   512 * 1024,  // Old pointer space.
   128 * 1024,  // Old data space.
